Aggregate builtin overloads into one builtin
